feature/feature tomorrow2

+++ user/inc/Sdcard.h
@@ -7,6 +7,6 @@ extern uint8 image_read_buffer[MT9V03X_H][MT9V03X_W];
 
 void sdcardinit();
 void sdcard_read();
-void sdcard_write(uint8 *image_write_buffer, uint16 size);
+void sdcard_write(int16 *image_write_buffer, uint16 size);
 
 #endif /* CODE_SDCARD_H_ */
